+++ CHANGELOG.md
@@ -9,6 +9,7 @@ in 1.x and 0.x versions.
   * Fix settings storage
   * Change startup file structure
   * Hide cutlist group's raw dimensions if no length or width increase is defined
+  * Added persistant material UUID mecanism to improve options storage
   
 * 1.5.0 (2018-06-25)