+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 This changelog references the relevant changes (bug and security fixes) done
 in 1.x and 0.x versions.
 
-* 1.5.1 (DEV)
+* 1.5.1 (2018-07-18)
 
   * Fix settings storage
   * Change startup file structure